move RPC-hungry tests to idle file for now
